smart_scout/requirements.txt
requests
beautifulsoup4
openai
python-dotenv
colorama
pyyaml
lxml
